test(amd): extract shared GSM8K helpers into gsm8k_completion_kit
